+++ internal/dataplane/kong_client_test.go
@@ -189,7 +189,10 @@ func newMockUpdateStrategyResolver(t *testing.T) *mockUpdateStrategyResolver {
 	}
 }
 
-func (f *mockUpdateStrategyResolver) ResolveUpdateStrategy(c sendconfig.UpdateClient) sendconfig.UpdateStrategy {
+func (f *mockUpdateStrategyResolver) ResolveUpdateStrategy(
+	c sendconfig.UpdateClient,
+	_ *diagnostics.ClientDiagnostic,
+) sendconfig.UpdateStrategy {
 	f.lock.Lock()
 	defer f.lock.Unlock()
 
@@ -952,7 +955,7 @@ func setupTestKongClient(
 ) *KongClient {
 	logger := zapr.NewLogger(zap.NewNop())
 	timeout := time.Second
-	diagnostic := diagnostics.ConfigDumpDiagnostic{}
+	diagnostic := diagnostics.ClientDiagnostic{}
 	config := sendconfig.Config{
 		SanitizeKonnectConfigDumps: true,
 	}
@@ -1247,7 +1250,7 @@ func TestKongClient_FallbackConfiguration_SuccessfulRecovery(t *testing.T) {
 			kongClient, err := NewKongClient(
 				zapr.NewLogger(zap.NewNop()),
 				time.Second,
-				diagnostics.ConfigDumpDiagnostic{
+				diagnostics.ClientDiagnostic{
 					Configs: diagnosticsCh,
 				},
 				sendconfig.Config{
@@ -1384,7 +1387,7 @@ func TestKongClient_FallbackConfiguration_SkipsUpdateWhenInSync(t *testing.T) {
 	kongClient, err := NewKongClient(
 		zapr.NewLogger(zap.NewNop()),
 		time.Second,
-		diagnostics.ConfigDumpDiagnostic{
+		diagnostics.ClientDiagnostic{
 			Configs: diagnosticsCh,
 		},
 		sendconfig.Config{
@@ -1533,7 +1536,7 @@ func TestKongClient_FallbackConfiguration_FailedRecovery(t *testing.T) {
 	kongClient, err := NewKongClient(
 		zapr.NewLogger(zap.NewNop()),
 		time.Second,
-		diagnostics.ConfigDumpDiagnostic{
+		diagnostics.ClientDiagnostic{
 			Configs: diagnosticsCh,
 		},
 		sendconfig.Config{
@@ -1646,7 +1649,7 @@ func TestKongClient_LastValidCacheSnapshot(t *testing.T) {
 			kongClient, err := NewKongClient(
 				zapr.NewLogger(zap.NewNop()),
 				time.Second,
-				diagnostics.ConfigDumpDiagnostic{},
+				diagnostics.ClientDiagnostic{},
 				sendconfig.Config{
 					FallbackConfiguration:         tc.fallbackConfigurationFeatureEnabled,
 					UseLastValidConfigForFallback: tc.useLastValidConfigForFallbackEnabled,
@@ -1734,7 +1737,7 @@ func TestKongClient_ConfigDumpSanitization(t *testing.T) {
 	for _, tc := range testCases {
 		t.Run(tc.name, func(t *testing.T) {
 			diagnosticsCh := make(chan diagnostics.ConfigDump, 1) // make it buffered to avoid blocking
-			kongClient.diagnostic = diagnostics.ConfigDumpDiagnostic{
+			kongClient.diagnostic = diagnostics.ClientDiagnostic{
 				Configs:               diagnosticsCh,
 				DumpsIncludeSensitive: tc.dumpsIncludeSensitive,
 			}
@@ -1869,7 +1872,7 @@ func TestKongClient_RecoveringFromGatewaySyncError(t *testing.T) {
 			kongClient, err := NewKongClient(
 				zapr.NewLogger(zap.NewNop()),
 				time.Second,
-				diagnostics.ConfigDumpDiagnostic{},
+				diagnostics.ClientDiagnostic{},
 				sendconfig.Config{
 					FallbackConfiguration: true,
 				},