+++ internal/diagnostics/server.go
@@ -26,13 +26,19 @@ const (
 	// variable) but do want a small amount of leeway to account for goroutine scheduling, so it
 	// is not zero.
 	diagnosticConfigBufferDepth = 3
+
+	// diffHistorySize is the number of diffs to keep in history.
+	diffHistorySize = 5
+
+	// diffHashQuery is the query string key for requesting a specific hash's diff.
+	diffHashQuery = "hash"
 )
 
 // Server is an HTTP server running exposing the pprof profiling tool, and processing diagnostic dumps of Kong configurations.
 type Server struct {
 	logger           logr.Logger
 	profilingEnabled bool
-	configDumps      ConfigDumpDiagnostic
+	clientDiagnostic ClientDiagnostic
 
 	lastSuccessfulConfigDump file.Content
 	lastSuccessHash          string
@@ -43,8 +49,11 @@ type Server struct {
 
 	currentFallbackCacheMetadata *fallback.GeneratedCacheMetadata
 
+	diffs diffMap
+
 	configLock   *sync.RWMutex
 	fallbackLock *sync.RWMutex
+	diffLock     *sync.RWMutex
 }
 
 // ServerConfig contains configuration for the diagnostics server.
@@ -66,13 +75,16 @@ func NewServer(logger logr.Logger, cfg ServerConfig) Server {
 		profilingEnabled: cfg.ProfilingEnabled,
 		configLock:       &sync.RWMutex{},
 		fallbackLock:     &sync.RWMutex{},
+		diffLock:         &sync.RWMutex{},
+		diffs:            newDiffMap(diffHistorySize),
 	}
 
 	if cfg.ConfigDumpsEnabled {
-		s.configDumps = ConfigDumpDiagnostic{
+		s.clientDiagnostic = ClientDiagnostic{
 			DumpsIncludeSensitive: cfg.DumpSensitiveConfig,
 			Configs:               make(chan ConfigDump, diagnosticConfigBufferDepth),
 			FallbackCacheMetadata: make(chan fallback.GeneratedCacheMetadata, diagnosticConfigBufferDepth),
+			Diffs:                 make(chan ConfigDiff, diagnosticConfigBufferDepth),
 		}
 	}
 
@@ -81,14 +93,14 @@ func NewServer(logger logr.Logger, cfg ServerConfig) Server {
 
 // ConfigDumps returns an object allowing dumping succeeded and failed configuration updates.
 // It will return a zero value of the type in case the config dumps are not enabled.
-func (s *Server) ConfigDumps() ConfigDumpDiagnostic {
-	return s.configDumps
+func (s *Server) ConfigDumps() ClientDiagnostic {
+	return s.clientDiagnostic
 }
 
 // Listen starts up the HTTP server and blocks until ctx expires.
 func (s *Server) Listen(ctx context.Context, port int) error {
 	mux := http.NewServeMux()
-	if s.configDumps != (ConfigDumpDiagnostic{}) {
+	if s.clientDiagnostic != (ClientDiagnostic{}) {
 		s.installConfigDebugHandlers(mux)
 	}
 	if s.profilingEnabled {
@@ -102,7 +114,7 @@ func (s *Server) Listen(ctx context.Context, port int) error {
 	}
 	errChan := make(chan error)
 
-	go s.receiveConfig(ctx)
+	go s.receiveDiagnostics(ctx)
 
 	go func() {
 		err := httpServer.ListenAndServe()
@@ -125,20 +137,22 @@ func (s *Server) Listen(ctx context.Context, port int) error {
 	}
 }
 
-// receiveConfig watches the config update channel.
-func (s *Server) receiveConfig(ctx context.Context) {
+// receiveDiagnostics watches the diagnostic update channels.
+func (s *Server) receiveDiagnostics(ctx context.Context) {
 	for {
 		select {
-		case dump := <-s.configDumps.Configs:
+		case dump := <-s.clientDiagnostic.Configs:
 			s.onConfigDump(dump)
-		case meta := <-s.configDumps.FallbackCacheMetadata:
+		case meta := <-s.clientDiagnostic.FallbackCacheMetadata:
 			s.onFallbackCacheMetadata(meta)
+		case diff := <-s.clientDiagnostic.Diffs:
+			s.onDiff(diff)
 		case <-ctx.Done():
 			if err := ctx.Err(); err != nil && !errors.Is(err, context.Canceled) {
-				s.logger.Error(err, "Shutting down diagnostic config collection: context completed with error")
+				s.logger.Error(err, "Shutting down diagnostic collection: context completed with error")
 				return
 			}
-			s.logger.V(logging.InfoLevel).Info("Shutting down diagnostic config collection: context completed")
+			s.logger.V(logging.InfoLevel).Info("Shutting down diagnostic collection: context completed")
 			return
 		}
 	}
@@ -174,6 +188,12 @@ func (s *Server) onFallbackCacheMetadata(meta fallback.GeneratedCacheMetadata) {
 	s.currentFallbackCacheMetadata = &meta
 }
 
+func (s *Server) onDiff(diff ConfigDiff) {
+	s.diffLock.Lock()
+	defer s.diffLock.Unlock()
+	s.diffs.Update(diff)
+}
+
 // installProfilingHandlers adds the Profiling webservice to the given mux.
 func installProfilingHandlers(mux *http.ServeMux) {
 	mux.HandleFunc("/debug/pprof", redirectTo("/debug/pprof/"))
@@ -195,6 +215,7 @@ func (s *Server) installConfigDebugHandlers(mux *http.ServeMux) {
 	mux.HandleFunc("/debug/config/failed", s.handleLastFailedConfig)
 	mux.HandleFunc("/debug/config/fallback", s.handleCurrentFallback)
 	mux.HandleFunc("/debug/config/raw-error", s.handleLastErrBody)
+	mux.HandleFunc("/debug/config/diff-report", s.handleDiffReport)
 }
 
 // redirectTo redirects request to a certain destination.
@@ -249,6 +270,68 @@ func (s *Server) handleLastErrBody(rw http.ResponseWriter, _ *http.Request) {
 		raw = []byte("No raw error body available.\n")
 	}
 	if _, err := rw.Write(raw); err != nil {
+		rw.WriteHeader(http.StatusInternalServerError)
+	}
+}
+
+func (s *Server) handleDiffReport(rw http.ResponseWriter, r *http.Request) {
+	rw.Header().Set("Content-Type", "application/json")
+	s.diffLock.RLock()
+	defer s.diffLock.RUnlock()
+
+	// GDR has no notion of sensitive data, so its raw diffs will include credentials and certificates when they
+	// change. We could make this fancier by walking through the entity types to exclude them if sensitive is not
+	// enabled, but would need to maintain a list of such types. Filter would probably happen on the producer (DB
+	// update strategy) side, since that's where we currently filter for the dump.
+	if !s.clientDiagnostic.DumpsIncludeSensitive {
+		if err := json.NewEncoder(rw).Encode(DiffResponse{
+			Message: "diffs include sensitive data: set CONTROLLER_DUMP_SENSITIVE_CONFIG=true in environment to enable",
+		}); err == nil {
+			rw.WriteHeader(http.StatusNotFound)
+		} else {
+			rw.WriteHeader(http.StatusInternalServerError)
+		}
+		return
+	}
+
+	if s.diffs.Len() == 0 {
+		if err := json.NewEncoder(rw).Encode(DiffResponse{
+			Message: "no diffs available",
+		}); err == nil {
+			rw.WriteHeader(http.StatusOK)
+		} else {
+			rw.WriteHeader(http.StatusInternalServerError)
+		}
+		return
+	}
+
+	var requestedHash string
+	var message string
+	requestedHashQuery := r.URL.Query()[diffHashQuery]
+	if len(requestedHashQuery) == 0 {
+		requestedHash = s.diffs.Latest()
+	} else {
+		if len(requestedHashQuery) > 1 {
+			message = "this endpoint does not support requesting multiple diffs, using the first hash provided"
+		}
+		requestedHash = requestedHashQuery[0]
+	}
+
+	requested, err := s.diffs.ByHash(requestedHash)
+	if err != nil {
+		message = err.Error()
+		rw.WriteHeader(http.StatusNotFound)
+	}
+
+	response := DiffResponse{
+		Message:    message,
+		ConfigHash: requestedHash,
+		Timestamp:  s.diffs.TimeByHash(requestedHash),
+		Diffs:      requested,
+		Available:  s.diffs.Available(),
+	}
+
+	if err := json.NewEncoder(rw).Encode(response); err != nil {
 		rw.WriteHeader(http.StatusInternalServerError)
 	}
 }