+++ internal/konnect/config_synchronizer.go
@@ -128,6 +128,7 @@ func (s *ConfigSynchronizer) uploadConfig(ctx context.Context, client *adminapi.
 		s.prometheusMetrics,
 		s.updateStrategyResolver,
 		s.configChangeDetector,
+		nil,
 		isFallback,
 	)
 	if err != nil {