+++ internal/manager/run.go
@@ -50,7 +50,7 @@ import (
 func Run(
 	ctx context.Context,
 	c *Config,
-	diagnostic diagnostics.ConfigDumpDiagnostic,
+	diagnostic diagnostics.ClientDiagnostic,
 	logger logr.Logger,
 ) error {
 	setupLog := ctrl.LoggerFrom(ctx).WithName("setup")